Consolidate EdgeProvider client logic
